+++ src/utility/util.cpp
@@ -1,4 +1,3 @@
-#include <bigint/BigIntegerLibrary.hh>
 #include "utility/util.h"
 
 namespace jub {
@@ -67,7 +66,7 @@ std::string DecStringToHexString(std::string str) {
     return bigUnsignedToString(stringToBigUnsigned(str.c_str()), 16);
 }
 
-std::string numberToHexString(unsigned int val) {
+std::string numberToHexString(unsigned long val) {
 
     return bigUnsignedToString(val, 16);
 }
